+++ examples/ridgemtask_classification.py
@@ -16,7 +16,7 @@ X_train, X_test, y_train, y_test = train_test_split(Z, t, test_size=0.2)
 
 print(Z.shape)
 
-fit_obj = ns.Ridge2ClassifierMtask(n_hidden_features=np.int(9.83730469e+01), 
+fit_obj = ns.Ridge2MultitaskClassifier(n_hidden_features=np.int(9.83730469e+01), 
                                    dropout=4.31054687e-01, 
                                    n_clusters=np.int(1.71484375e+00),
                                    lambda1=1.24023438e+01, lambda2=7.30263672e+03)
@@ -43,7 +43,7 @@ t = wine.target
 np.random.seed(123)
 Z_train, Z_test, y_train, y_test = train_test_split(Z, t, test_size=0.2)
 
-fit_obj = ns.Ridge2ClassifierMtask(n_hidden_features=15,
+fit_obj = ns.Ridge2MultitaskClassifier(n_hidden_features=15,
                                   dropout=0.1, n_clusters=3, 
                                   type_clust="gmm")
 
@@ -66,7 +66,7 @@ np.random.seed(123)
 Z_train, Z_test, y_train, y_test = train_test_split(Z, t, test_size=0.2)
 
 
-fit_obj = ns.Ridge2ClassifierMtask(n_hidden_features=10,
+fit_obj = ns.Ridge2MultitaskClassifier(n_hidden_features=10,
                                   dropout=0.1, n_clusters=2)
 
 start = time()
@@ -84,7 +84,7 @@ X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2,
                                                     random_state=351452)
 
 
-fit_obj = ns.Ridge2ClassifierMtask(n_hidden_features=5,
+fit_obj = ns.Ridge2MultitaskClassifier(n_hidden_features=5,
                                   dropout=0.1, n_clusters=3, 
                                   type_clust="gmm")
 
@@ -106,7 +106,7 @@ y = digits.target
 X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2,
                                                     random_state=123)
 
-fit_obj = ns.Ridge2ClassifierMtask(n_hidden_features=25,
+fit_obj = ns.Ridge2MultitaskClassifier(n_hidden_features=25,
                                   dropout=0.1, n_clusters=3, 
                                   type_clust="gmm")