+++ nnetsauce/ridge2/__init__.py
@@ -1,5 +1,5 @@
 from .ridge2Classifier import Ridge2Classifier
 from .ridge2Regressor import Ridge2Regressor
-from .ridge2ClassifierMtask import Ridge2ClassifierMtask
+from .ridge2MultitaskClassifier import Ridge2MultitaskClassifier
 
-__all__ = ["Ridge2Classifier", "Ridge2Regressor", "Ridge2ClassifierMtask"]
+__all__ = ["Ridge2Classifier", "Ridge2Regressor", "Ridge2MultitaskClassifier"]