+++ nnetsauce/ridge2/ridge2MultitaskClassifier.py
@@ -1,25 +1,24 @@
+
 # Authors: Thierry Moudiki
 #
 # License: BSD 3
 
 import numpy as np
-import pickle
+from scipy.optimize import minimize
 import sklearn.metrics as skm2
+from .ridge2 import Ridge2
 from ..utils import matrixops as mo
 from ..utils import misc as mx
-from ..base import Base
-from scipy.special import expit
 from sklearn.base import ClassifierMixin
+from scipy.special import logsumexp
+from scipy.linalg import pinv
 
 
-class MtaskClassifier(Base, ClassifierMixin):
-    """Multitask Classification model based on regression models, with shared covariates
+class Ridge2MultitaskClassifier(Ridge2, ClassifierMixin):
+    """Multitask Ridge classification with 2 regularization parameters
     
        Parameters
        ----------
-       obj: object
-           any object (must be a regression model) containing a method fit (obj.fit()) 
-           and a method predict (obj.predict())
        n_hidden_features: int
            number of nodes in the hidden layer
        activation_name: str
@@ -35,9 +34,6 @@ class MtaskClassifier(Base, ClassifierMixin):
        dropout: float
            regularization parameter; (random) percentage of nodes dropped out 
            of the training
-       direct_link: boolean
-           indicates if the original predictors are included (True) in model's 
-           fitting or not (False)
        n_clusters: int
            number of clusters for 'kmeans' or 'gmm' clustering (could be 0: 
                no clustering)
@@ -51,10 +47,10 @@ class MtaskClassifier(Base, ClassifierMixin):
            scaling methods for inputs, hidden layer, and clustering respectively
            (and when relevant). 
            Currently available: standardization ('std') or MinMax scaling ('minmax')
-       col_sample: float
-           percentage of covariates randomly chosen for training   
-       row_sample: float
-           percentage of rows chosen for training, by stratified bootstrapping    
+       lambda1: float
+           regularization parameter on direct link
+       lambda2: float
+           regularization parameter on hidden layer
        seed: int 
            reproducibility seed for nodes_sim=='uniform'
     """
@@ -63,20 +59,18 @@ class MtaskClassifier(Base, ClassifierMixin):
 
     def __init__(
         self,
-        obj,
         n_hidden_features=5,
         activation_name="relu",
         a=0.01,
         nodes_sim="sobol",
         bias=True,
         dropout=0,
-        direct_link=True,
         n_clusters=2,
         cluster_encode=True,
         type_clust="kmeans",
         type_scaling=("std", "std", "std"),
-        col_sample=1,
-        row_sample=1,
+        lambda1=0.1,
+        lambda2=0.1,
         seed=123,
     ):
 
@@ -87,22 +81,22 @@ class MtaskClassifier(Base, ClassifierMixin):
             nodes_sim=nodes_sim,
             bias=bias,
             dropout=dropout,
-            direct_link=direct_link,
             n_clusters=n_clusters,
             cluster_encode=cluster_encode,
             type_clust=type_clust,
             type_scaling=type_scaling,
-            col_sample=col_sample,
-            row_sample=row_sample,
+            col_sample=1,
+            row_sample=1,
+            lambda1=lambda1,
+            lambda2=lambda2,
             seed=seed,
         )
 
-        self.type_fit = "classification"
-        self.obj = obj
-        self.fit_objs = {}
+        self.type_fit = "regression"
 
-    def fit(self, X, y, sample_weight=None, **kwargs):
-        """Fit MtaskClassifier to training data (X, y).
+    def fit(self, X, y, **kwargs):
+        """Fit Ridge model to training data (X, y).           
+
         
         Parameters
         ----------
@@ -125,17 +119,41 @@ class MtaskClassifier(Base, ClassifierMixin):
 
         output_y, scaled_Z = self.cook_training_set(y=y, X=X, **kwargs)
 
+        n_X, p_X = X.shape
+        n_Z, p_Z = scaled_Z.shape
+
         self.n_classes = len(np.unique(y))
 
         # multitask response
         Y = mo.one_hot_encode2(output_y, self.n_classes)
 
-        # if sample_weight is None:
-        for i in range(self.n_classes):
+        if self.n_clusters > 0:
+            if self.encode_clusters == True:
+                n_features = p_X + self.n_clusters
+            else:
+                n_features = p_X + 1
+        else:
+            n_features = p_X
 
-            self.fit_objs[i] = pickle.loads(
-                pickle.dumps(self.obj.fit(scaled_Z, Y[:, i], **kwargs), -1)
-            )
+        X_ = scaled_Z[:, 0:n_features]
+        Phi_X_ = scaled_Z[:, n_features:p_Z]
+
+        B = mo.crossprod(X_) + self.lambda1 * np.diag(np.repeat(1, X_.shape[1]))
+        C = mo.crossprod(Phi_X_, X_)
+        D = mo.crossprod(Phi_X_) + self.lambda2 * np.diag(
+            np.repeat(1, Phi_X_.shape[1])
+        )
+        B_inv = pinv(B)
+        W = np.dot(C, B_inv)
+        S_mat = D - mo.tcrossprod(W, C)
+        S_inv = pinv(S_mat)
+        Y2 = np.dot(S_inv, W)
+        inv = mo.rbind(
+            mo.cbind(B_inv + mo.crossprod(W, Y2), -np.transpose(Y2)),
+            mo.cbind(-Y2, S_inv),
+        )
+        
+        self.beta = np.dot(inv, mo.crossprod(scaled_Z, Y))
 
         return self
 
@@ -175,14 +193,9 @@ class MtaskClassifier(Base, ClassifierMixin):
         probability estimates for test data: {array-like}        
         """
 
-        shape_X = X.shape
+        if len(X.shape) == 1:
 
-        probs = np.zeros((shape_X[0], self.n_classes))
-
-        if len(shape_X) == 1:
-
-            n_features = shape_X[0]
-
+            n_features = X.shape[0]
             new_X = mo.rbind(
                 X.reshape(1, n_features),
                 np.ones(n_features).reshape(1, n_features),
@@ -190,47 +203,18 @@ class MtaskClassifier(Base, ClassifierMixin):
 
             Z = self.cook_test_set(new_X, **kwargs)
 
-            # loop on all the classes
-            for i in range(self.n_classes):
-
-                probs[:, i] = self.fit_objs[i].predict(Z, **kwargs)[0]
-
         else:
 
             Z = self.cook_test_set(X, **kwargs)
 
-            # loop on all the classes
-            for i in range(self.n_classes):
+        ZB = mo.safe_sparse_dot(Z, self.beta)
 
-                probs[:, i] = self.fit_objs[i].predict(Z, **kwargs)
+        exp_ZB = np.exp(ZB)
 
-        expit_raw_probs = expit(probs)
-
-        return expit_raw_probs / expit_raw_probs.sum(axis=1)[:, None]
+        return exp_ZB / exp_ZB.sum(axis=1)[:, None]
 
     def score(self, X, y, scoring=None, **kwargs):
-        """ Score the model on test set features X and response y. 
-
-        Parameters
-        ----------
-        X: {array-like}, shape = [n_samples, n_features]
-            Training vectors, where n_samples is the number 
-            of samples and n_features is the number of features
-
-        y: array-like, shape = [n_samples]
-            Target values
-
-        scoring: str
-            must be in ('explained_variance', 'neg_mean_absolute_error', \
-                        'neg_mean_squared_error', 'neg_mean_squared_log_error', \
-                        'neg_median_absolute_error', 'r2')
-        
-        **kwargs: additional parameters to be passed to scoring functions
-               
-        Returns
-        -------
-        model scores: {array-like}
-        """
+        """ Score the model on test set covariates X and response y. """
 
         preds = self.predict(X)